+++ src/image_write.h
@@ -9,7 +9,7 @@
 #include "swath.h"
 
 void _write_tif(
-    std::vector<float>& img_data,
+    std::vector<double>& img_data,
     const int& rows,
     const int& cols,
     const std::string out_filename
